Refactor DDS Fuzz For Reusability
