+++ packages/dds/sequence/src/test/intervalTestUtils.ts
@@ -6,7 +6,7 @@
 import { strict as assert } from "assert";
 
 import { isObject } from "@fluidframework/core-utils/internal";
-import { isFluidHandle } from "@fluidframework/runtime-utils/internal";
+import { isFluidHandle, toFluidHandleInternal } from "@fluidframework/runtime-utils/internal";
 import { MockContainerRuntimeForReconnection } from "@fluidframework/test-runtime-utils/internal";
 
 import { IIntervalCollection } from "../intervalCollection.js";
@@ -129,8 +129,14 @@ async function assertPropertiesEqual(a: SharedString, b: SharedString): Promise<
 		for (const key of aKeys.concat(bKeys)) {
 			const aVal: unknown = aProps[key];
 			const bVal: unknown = bProps[key];
-			const aHandle = isObject(aVal) && isFluidHandle(aVal) ? await aVal.get() : aVal;
-			const bHandle = isObject(bVal) && isFluidHandle(bVal) ? await bVal.get() : bVal;
+			const aHandle =
+				isObject(aVal) && isFluidHandle(aVal)
+					? toFluidHandleInternal(aVal).absolutePath
+					: aVal;
+			const bHandle =
+				isObject(bVal) && isFluidHandle(bVal)
+					? toFluidHandleInternal(bVal).absolutePath
+					: bVal;
 			assert.deepEqual(
 				aHandle,
 				bHandle,
