+++ packages/dds/test-dds-utils/src/ddsFuzzHarness.ts
@@ -5,27 +5,32 @@
 
 import { strict as assert } from "node:assert";
 import { mkdirSync, readFileSync } from "node:fs";
-import path from "node:path";
 
 import { TypedEventEmitter } from "@fluid-internal/client-utils";
 import type {
 	AsyncGenerator,
 	AsyncReducer,
 	BaseFuzzTestState,
+	BaseOperation,
 	IRandom,
+	MinimizationTransform,
 	SaveDestination,
 	SaveInfo,
 } from "@fluid-private/stochastic-test-utils";
 import {
 	ExitBehavior,
-	StressMode,
+	FuzzTestMinimizer,
 	asyncGeneratorFromArray,
 	chainAsync,
 	createFuzzDescribe,
 	createWeightedAsyncGenerator,
 	defaultOptions,
 	done,
+	generateTestSeeds,
+	getSaveDirectory,
+	getSaveInfo,
 	interleaveAsync,
+	isOperationType,
 	makeRandom,
 	performFuzzActionsAsync,
 	saveOpsToFile,
@@ -59,13 +64,6 @@ import {
 	hasStashData,
 } from "./clientLoading.js";
 import { DDSFuzzHandle } from "./ddsFuzzHandle.js";
-import type { MinimizationTransform } from "./minification.js";
-import { FuzzTestMinimizer } from "./minification.js";
-
-const isOperationType = <O extends BaseOperation>(
-	type: O["type"],
-	op: BaseOperation,
-): op is O => op.type === type;
 
 /**
  * @internal
@@ -102,13 +100,6 @@ export interface DDSFuzzTestState<TChannelFactory extends IChannelFactory>
  */
 export interface ClientSpec {
 	clientId: string;
-}
-
-/**
- * @internal
- */
-export interface BaseOperation {
-	type: number | string;
 }
 
 /**
@@ -180,37 +171,6 @@ export interface AddClient {
 export interface Synchronize {
 	type: "synchronize";
 	clients?: string[];
-}
-
-/**
- * @internal
- */
-interface HasWorkloadName {
-	workloadName: string;
-}
-
-function getSaveDirectory(directory: string, model: HasWorkloadName): string {
-	const workloadFriendly = model.workloadName.replace(/[\s_]+/g, "-").toLowerCase();
-	return path.join(directory, workloadFriendly);
-}
-
-function getSavePath(directory: string, model: HasWorkloadName, seed: number): string {
-	return path.join(getSaveDirectory(directory, model), `${seed}.json`);
-}
-
-function getSaveInfo(
-	model: HasWorkloadName,
-	options: DDSFuzzSuiteOptions,
-	seed: number,
-): SaveInfo {
-	return {
-		saveOnFailure: options.saveFailures
-			? { path: getSavePath(options.saveFailures.directory, model, seed) }
-			: false,
-		saveOnSuccess: options.saveSuccesses
-			? { path: getSavePath(options.saveSuccesses.directory, model, seed) }
-			: false,
-	};
 }
 
 /**
@@ -297,7 +257,7 @@ export interface DDSFuzzModel<
 
 	/**
 	 * An array of transforms used during fuzz test minimization to reduce test
-	 * cases. See {@link MinimizationTransform} for additional context.
+	 * cases. See {@link @fluid-private/stochastic-test-utils#MinimizationTransform} for additional context.
 	 *
 	 * If no transforms are supplied, minimization will still occur, but the
 	 * contents of the operations will remain unchanged.
@@ -529,7 +489,7 @@ export interface DDSFuzzSuiteOptions {
 	 * useful if the model being tested defines {@link DDSFuzzModel.minimizationTransforms}.
 	 *
 	 * It can also add a couple seconds of overhead per failing
-	 * test case. See {@link MinimizationTransform} for additional context.
+	 * test case. See {@link @fluid-private/stochastic-test-utils#MinimizationTransform} for additional context.
 	 */
 	skipMinimization?: boolean;
 
@@ -1526,7 +1486,13 @@ function runTest<TChannelFactory extends IChannelFactory, TOperation extends Bas
 				throw error;
 			}
 			const operations = JSON.parse(file.toString()) as TOperation[];
-			const minimizer = new FuzzTestMinimizer(model, options, operations, seed, saveInfo, 3);
+			const minimizer = new FuzzTestMinimizer(
+				model.minimizationTransforms,
+				operations,
+				saveInfo,
+				async (generator) => replayTest(model, seed, generator, saveInfo, options),
+				3,
+			);
 
 			const minimized = await minimizer.minimize();
 			await saveOpsToFile(savePath, minimized);
@@ -1566,7 +1532,7 @@ export async function replayTest<
 >(
 	ddsModel: DDSFuzzModel<TChannelFactory, TOperation>,
 	seed: number,
-	operations: TOperation[],
+	generator: AsyncGenerator<TOperation, unknown>,
 	saveInfo?: SaveInfo,
 	providedOptions?: Partial<DDSFuzzSuiteOptions>,
 ): Promise<void> {
@@ -1582,36 +1548,10 @@ export async function replayTest<
 	const model = {
 		..._model,
 		// We lose some type safety here because the options interface isn't generic
-		generatorFactory: (): AsyncGenerator<TOperation, unknown> =>
-			asyncGeneratorFromArray(operations),
+		generatorFactory: (): AsyncGenerator<TOperation, unknown> => generator,
 	};
 
 	await runTestForSeed(model, options, seed, saveInfo);
-}
-
-export function generateTestSeeds(testCount: number, stressMode: StressMode): number[] {
-	switch (stressMode) {
-		case StressMode.Short:
-		case StressMode.Normal: {
-			// Deterministic, fixed seeds
-			return Array.from({ length: testCount }, (_, i) => i);
-		}
-
-		case StressMode.Long: {
-			// Non-deterministic, random seeds
-			const random = makeRandom();
-			const longModeFactor = 2;
-			const initialSeed = random.integer(
-				0,
-				Number.MAX_SAFE_INTEGER - longModeFactor * testCount,
-			);
-			return Array.from({ length: testCount * longModeFactor }, (_, i) => initialSeed + i);
-		}
-
-		default: {
-			throw new Error(`Unsupported stress mode: ${stressMode}`);
-		}
-	}
 }
 
 /**
