+++ packages/dds/test-dds-utils/src/index.ts
@@ -7,7 +7,6 @@ export type { IGCTestProvider } from "./gcTestRunner.js";
 export { runGCTests } from "./gcTestRunner.js";
 export type {
 	AddClient,
-	BaseOperation,
 	ChangeConnectionState,
 	ClientSpec,
 	DDSFuzzModel,
@@ -24,5 +23,4 @@ export {
 } from "./ddsFuzzHarness.js";
 export type { ISnapshotSuite } from "./ddsSnapshotHarness.js";
 export { createSnapshotSuite } from "./ddsSnapshotHarness.js";
-export type { MinimizationTransform } from "./minification.js";
 export type { Client, FuzzSerializedIdCompressor } from "./clientLoading.js";
