+++ packages/test/stochastic-test-utils/package.json
@@ -94,7 +94,8 @@
 	},
 	"dependencies": {
 		"@fluidframework/core-utils": "workspace:~",
-		"best-random": "^1.0.0"
+		"best-random": "^1.0.0",
+		"path-browserify": "^1.0.1"
 	},
 	"devDependencies": {
 		"@arethetypeswrong/cli": "^0.17.1",
@@ -108,6 +109,7 @@
 		"@microsoft/api-extractor": "7.47.8",
 		"@types/mocha": "^10.0.10",
 		"@types/node": "^18.19.0",
+		"@types/path-browserify": "^1.0.0",
 		"c8": "^8.0.1",
 		"concurrently": "^8.2.1",
 		"copyfiles": "^2.4.1",
