+++ packages/test/stochastic-test-utils/src/combineReducers.ts
@@ -10,8 +10,23 @@ import { AsyncReducer, BaseFuzzTestState, Reducer } from "./types.js";
 /**
  * @internal
  */
+export interface BaseOperation {
+	type: number | string;
+}
+
+/**
+ * @internal
+ */
+export const isOperationType = <O extends BaseOperation>(
+	type: O["type"],
+	op: BaseOperation,
+): op is O => op.type === type;
+
+/**
+ * @internal
+ */
 export function combineReducers<
-	TOperation extends { type: string | number },
+	TOperation extends BaseOperation,
 	TState extends BaseFuzzTestState,
 >(
 	reducerMap: {
@@ -33,7 +48,7 @@ export function combineReducers<
  * @internal
  */
 export function combineReducersAsync<
-	TOperation extends { type: string | number },
+	TOperation extends BaseOperation,
 	TState extends BaseFuzzTestState,
 >(
 	reducerMap: {
@@ -50,3 +65,13 @@ export function combineReducersAsync<
 		return newState;
 	};
 }
+
+/**
+ * Some reducers require preconditions be met which are validated by their generator.
+ * The validation can be lost if the generator is not run.
+ * The primary case where this happens is during minimization. If a reducer detects this
+ * problem, they can throw this error type, and minimization will consider the current
+ * test invalid, rather than continuing to test invalid scenarios.
+ * @internal
+ */
+export class ReducerPreconditionError extends Error {}
